requirements.txt
streamlit
tensorflow
pillow
numpy
pandas
matplotlib
boto3